+++ rich/repr.py
@@ -1,3 +1,4 @@
+from functools import partial
 import inspect
 
 from typing import (
@@ -25,22 +26,22 @@ class ReprError(Exception):
 
 
 @overload
-def auto(cls: Type[T]) -> Type[T]:
+def auto(cls: Optional[T]) -> T:
     ...
 
 
 @overload
-def auto(*, angular: bool = False) -> Callable[[Type[T]], Type[T]]:
+def auto(*, angular: bool = False) -> Callable[[T], T]:
     ...
 
 
 def auto(
-    cls: Optional[Type[T]] = None, *, angular: bool = False
-) -> Union[Type[T], Callable[[Type[T]], Type[T]]]:
+    cls: Optional[T] = None, *, angular: Optional[bool] = None
+) -> Union[T, Callable[[T], T]]:
     """Class decorator to create __repr__ from __rich_repr__"""
 
-    def do_replace(cls: Type[T]) -> Type[T]:
-        def auto_repr(self: T) -> str:
+    def do_replace(cls: Type[T], angular: Optional[bool] = None) -> Type[T]:
+        def auto_repr(self: Type[T]) -> str:
             """Create repr string from __rich_repr__"""
             repr_str: List[str] = []
             append = repr_str.append
@@ -65,10 +66,10 @@ def auto(
             else:
                 return f"{self.__class__.__name__}({', '.join(repr_str)})"
 
-        def auto_rich_repr(self: T) -> RichReprResult:
+        def auto_rich_repr(self: Type[T]) -> RichReprResult:
             """Auto generate __rich_rep__ from signature of __init__"""
             try:
-                signature = inspect.signature(self.__init__)  # type: ignore
+                signature = inspect.signature(self.__init__)  ## type: ignore
                 for name, param in signature.parameters.items():
                     if param.kind == param.POSITIONAL_ONLY:
                         yield getattr(self, name)
@@ -88,20 +89,36 @@ def auto(
         if not hasattr(cls, "__rich_repr__"):
             auto_rich_repr.__doc__ = "Build a rich repr"
             cls.__rich_repr__ = auto_rich_repr  # type: ignore
-            cls.__rich_repr__.angular = angular  # type: ignore
 
         auto_repr.__doc__ = "Return repr(self)"
         cls.__repr__ = auto_repr  # type: ignore
+        if angular is not None:
+            cls.__rich_repr__.angular = angular  # type: ignore
         return cls
 
     if cls is None:
-        angular = angular
-        return do_replace
+        return partial(do_replace, angular=angular)  # type: ignore
     else:
-        return do_replace(cls)
+        return do_replace(cls, angular=angular)  # type: ignore
 
 
-rich_repr: Callable[[Type[T]], Type[T]] = auto
+@overload
+def rich_repr(cls: Optional[T]) -> T:
+    ...
+
+
+@overload
+def rich_repr(*, angular: bool = False) -> Callable[[T], T]:
+    ...
+
+
+def rich_repr(
+    cls: Optional[T] = None, *, angular: bool = False
+) -> Union[T, Callable[[T], T]]:
+    if cls is None:
+        return auto(angular=angular)
+    else:
+        return auto(cls)
 
 
 if __name__ == "__main__":