+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.5.0] - 2021-05-07
+
+### Fixed
+
+- Fixed Pandas objects not pretty printing https://github.com/willmcgugan/rich/issues/1305
+- Fixed https://github.com/willmcgugan/rich/issues/1256
+- Fixed typing with rich.repr.auto decorator
+- Fixed repr error formatting https://github.com/willmcgugan/rich/issues/1326
+
+### Added
+
+- Added new_line_start argument to Console.print
+- Added Segment.divide method
+- Added Segment.split_cells method
+- Added segment.SegmentLines class
+
 ## [10.4.0] - 2021-06-18
 
 ### Added